refactor tasks e comments

+++ to-do/src/main/java/com/example/todo/services/StatusService.java
@@ -69,6 +69,11 @@ public class StatusService {
     @Transactional
     public ResponseEntity<Object> deleteStatusById(Long id) {
 
+        if (taskRepository.existsByStatusId(id)) {
+
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("Status atrelados a tasks não podem ser apagados!");
+        }
+
         repository.deleteById(id);
 
         return ResponseEntity.status(HttpStatus.OK).build();
